+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add support for arithmetic operators
 * Add support for accessing properties using bracket notation
 * Update `CONTAINS` operator to support objects
+* Track campaign, group, and promo impressions
 
 ## 1.2.0 (2019-04-11)